# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My APIs",
        "contact": {
            "email": "weisse011@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "list tasks",
                "operationId": "listTasks",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "unexpected error"
                    }
                }
            },
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Create a task",
                "operationId": "createTask",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "new",
                            "enum": [
                                "new",
                                "processing",
                                "checking",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "executor",
                        "in": "query",
                        "description": "id of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "enum": [
                                1,
                                2,
                                3,
                                4,
                                5,
                                6,
                                7,
                                8,
                                9,
                                10
                            ]
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Task is successfuly created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "uprocessable content"
                    },
                    "500": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Get a Task",
                "description": "Returns a task based on a single ID",
                "operationId": "findTaskById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of task to fetch",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "show task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "unexpected error"
                    },
                    "404": {
                        "description": "not found task error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "tasks"
                ],
                "summary": "Delete a Task",
                "description": "deletes a single task based on the ID supplied",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of task to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "task deleted"
                    },
                    "500": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/api/tasks/{taskId}": {
            "put": {
                "tags": [
                    "tasks"
                ],
                "summary": "Update a task",
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "ID of task to update",
                        "required": true
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "new",
                            "enum": [
                                "new",
                                "processing",
                                "checking",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "executor",
                        "in": "query",
                        "description": "id of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "enum": [
                                1,
                                2,
                                3,
                                4,
                                5,
                                6,
                                7,
                                8,
                                9,
                                10
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "uprocessable content"
                    },
                    "500": {
                        "description": "unexpected error"
                    },
                    "404": {
                        "description": "not found task error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Task": {
                "required": [
                    "title",
                    "description",
                    "status",
                    "executor"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "description": "Task title",
                        "type": "string",
                        "example": "random title"
                    },
                    "description": {
                        "description": "Task description",
                        "type": "string",
                        "example": "random description"
                    },
                    "status": {
                        "description": "Task staus",
                        "type": "string",
                        "example": "new"
                    },
                    "executor": {
                        "description": "Task executor",
                        "type": "integer",
                        "example": "{ id:1, name: randomName }"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
        }
    }
}